Refactor and delete old and unused code
