+++ src/Altinn.App.Core/Internal/Process/IProcessNavigator.cs
@@ -1,8 +1,7 @@
-using Altinn.App.Core.Features;
 using Altinn.App.Core.Internal.Process.Elements.Base;
 using Altinn.Platform.Storage.Interface.Models;
 
-namespace Altinn.App.Core.Internal.Process.V2
+namespace Altinn.App.Core.Internal.Process
 {
     /// <summary>
     /// Interface used to descipt the process navigator
@@ -16,6 +15,6 @@ namespace Altinn.App.Core.Internal.Process.V2
         /// <param name="currentElement">Current process element id</param>
         /// <param name="action">Action performed</param>
         /// <returns>The next process task</returns>
-        public Task<ProcessElement> GetNextTask(Instance instance, string currentElement, string? action);
+        public Task<ProcessElement?> GetNextTask(Instance instance, string currentElement, string? action);
     }
 }
