--- src/Altinn.App.Core/Internal/Process/V2/IProcessEventDispatcher.cs
@@ -1,9 +1,0 @@
-using Altinn.Platform.Storage.Interface.Models;
-
-namespace Altinn.App.Core.Internal.Process.V2;
-
-public interface IProcessEventDispatcher
-{
-    Task<Instance> UpdateProcessAndDispatchEvents(Instance instance, Dictionary<string, string>? prefill, List<InstanceEvent> events);
-    Task RegisterEventWithEventsComponent(Instance instance);
-}
